Move BinarySensor to its own file
